Integrated Domain, Infrastructure and API layer - BankController Created

+++ src/SimpleBank.Domain/Contracts/IRepository.cs
@@ -4,7 +4,7 @@ namespace SimpleBank.Domain.Contracts
 {
     public interface IRepository<T> where T: Entity, IAggregateRoot
     {
-        Task<T?> GetById(int id);
+        Task<T?> GetById(long id);
         Task<IEnumerable<T>> List();
         Task<IEnumerable<T>> List(Expression<Func<T, bool>> predicate);
         Task<T> Add(T entity);
